added usercutoff and generate cut off

+++ payroll/src/main/java/org/ideyatech/payroll/entity/UserCutOff.java
@@ -107,7 +107,7 @@ public class UserCutOff {
 		this.nontaxable = nontaxable;
 	}
 
-	public double getNumberofdependents() {
+	public int getNumberofdependents() {
 		return numberofdependents;
 	}
